ref: Rename file name to match component name
